+++ src/redux/actions/bootstrap.ts
@@ -1,0 +1,5 @@
+import { createAction } from 'redux-actions'
+
+export const BOOTSTRAP = 'RESET_APP_STATE';
+export const bootstrapAction = createAction(BOOTSTRAP);
+export type BootStrapAction = ReturnType<typeof bootstrapAction>
